+++ src/db/migration/mod.rs
@@ -5,5 +5,6 @@ mod db_migration;
 mod migration_map;
 mod v0_12_1;
 mod v0_13_0;
+mod void_migration;
 
 pub use db_migration::DbMigration;